refactor(publish): isolate Obsidian image parsing
